Added logout page and integrated with login/register and home stub
Added home stub with logout button that redirects to logout confirmation page.
Integrated login/register page (which is broken as of 11/29/2019).
Changed connection code from localhost to Flip3 port #.

Please read the readme txt file for details on the URLs to access the site, and how to run your own instance of this site.